% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_n_rec.R
\name{adjust_N_recommendation}
\alias{adjust_N_recommendation}
\title{Adjust raster N recommendation}
\usage{
adjust_N_recommendation(n_rec, min_appl_rate = NULL, max_appl_rate = NULL)
}
\description{
Adjust raster N recommendation
}
