% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Calculates ndvi}
\usage{
calc_ndvi(NIR, Red)
}
\arguments{
\item{NIR}{The spectral reflectance in NIR region.}

\item{Red}{The spectral reflectance in red or visible region.}
}
\value{
ndvi
}
\description{
This function calculates ndvi as a function of near infrared) NIR and red channels.
}
\examples{
calc_ndvi(0.5, 0.08)
}
