% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{calc_sentera_ndvi}
\alias{calc_sentera_ndvi}
\title{Calculates NDVI using equation from Sentera}
\usage{
calc_sentera_ndvi(NIR, red)
}
\arguments{
\item{NIR}{The spectral reflectance in the NIR region.}

\item{red}{The spectral reflectance in red or visible region.}
}
\value{
NDVI
}
\description{
This function calculates NDVI as a function of NIR and red channels using
the equation provided by Sentera
}
\examples{
calc_sentera_ndvi(0.5, 0.08)
}
