% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_return_on_investment.R
\name{calculate_return_on_investment}
\alias{calculate_return_on_investment}
\title{Function for calculating the return on investment for variable rate nitrogen application}
\usage{
calculate_return_on_investment(yield_fertilized, yield_field, N_rate,
  grain_price, N_price)
}
\arguments{
\item{yield_uniform_N}{yield prediction with uniform rate nitrogen application}

\item{yield_no_N}{yield prediction without any nitrogen application}
}
\value{
The SBNRC adjusted Nitrogen rate
}
\description{
Function for calculating the return on investment for variable rate nitrogen application
}
\examples{

}
