% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBNRC_Algorithm.R
\name{fertilized_yield_potential}
\alias{fertilized_yield_potential}
\title{Calculate yield potential}
\usage{
fertilized_yield_potential(
  yp_field,
  yp_N_rich,
  N_rich_strip,
  farmer_practice,
  slope,
  intercept
)
}
\description{
Calculate yield potential
}
