% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_crs.R
\name{get_local_crs}
\alias{get_local_crs}
\title{Convert latitude and longitude to corresponding UTM projection}
\usage{
get_local_crs(lat, lon)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}
}
\description{
Convert latitude and longitude to corresponding UTM projection
}
