% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_active_ndvi.R
\name{map_active_ndvi}
\alias{map_active_ndvi}
\title{Interpolates active ndvi using regression kriging.}
\usage{
map_active_ndvi(pts, img)
}
\arguments{
\item{pts}{spdf of actively measured ndvi.}

\item{img}{raster layer of passive ndvi.}
}
\value{
image of the kriged values.
}
\description{
This function kriges active ndvi from actively measured ndvi points and passive ndvi values from image and returns the kriged values as an image.
The autokrige function from automap package is used to do the kriging.
}
