% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_image}
\alias{read_image}
\title{Converts multiband geotiff to ndvi image.}
\usage{
read_image(filename, red_band = 6, nir_band = 4)
}
\arguments{
\item{filename}{The tiff file with multiple bands.}
}
\value{
ndvi image
}
\description{
This function reads in a multiband geotiff file, calculates ndvi using calc_ndvi function, projects the raster into crs utm14, and returns ndvi image as a single layer.
}
\examples{
read_image('MULTIBAND.tiff')
}
