% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_n_rich.R
\name{read_n_rich_strip_farmer_practice}
\alias{read_n_rich_strip_farmer_practice}
\title{Reads shapefile of N Rich Strip and buffers for Farmer practice}
\usage{
read_n_rich_strip_farmer_practice(shapefile_path, active_ndvi)
}
\arguments{
\item{shapefile_path}{a length-one character vector that gives the file path
for the shapefile for the N rich strip}
}
\description{
Reads shapefile of N Rich Strip and buffers for Farmer practice
}
