% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_pts}
\alias{read_pts}
\title{Converts csv with crs latitude, longitude to spatial points data frame with crs Universal Transverse Mercator 14 (utm14).}
\usage{
read_pts(filename,
  t_srs = CRS("+proj=utm +zone=14 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"))
}
\arguments{
\item{filename}{The csv file with Normalized difference vegetation index (ndvi) data.}
}
\value{
The transformed points.
}
\description{
This function reads in a csv file with crs latitude and longitude, converts it to a spatial points data frame, and reprojects it to crs utm zone 14.
}
\examples{
read_pts('ndvi.csv')
}
