% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_sentera}
\alias{read_sentera}
\title{Converts a False-color geotiff to NDVI image using Sentera's formula}
\usage{
read_sentera(filename, red_band = 1, NIR_band = 3)
}
\arguments{
\item{filename}{The tiff file with multiple bands.}
}
\value{
ndvi image
}
\description{
This function reads in a false-color geotiff file, calculates NDVI
 using calc_sentera_ndvi() function and returns ndvi image as a single layer.
}
\examples{
read_sentera_rgb('FALSE_COLOR.tiff')
}
