% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zero_check.R
\name{read_zero_n_check_farmer_practice}
\alias{read_zero_n_check_farmer_practice}
\title{Reads shapefile of Zero Check Strip and buffers for Farmer practice}
\usage{
read_zero_n_check_farmer_practice(
  shapefile_path,
  active_ndvi,
  shapefile_crs = NULL
)
}
\arguments{
\item{shapefile_path}{a length-one character vector that gives the file path
for the shapefile for the N rich strip}

\item{active_ndvi}{a raster layer that contains actively mapped NDVI}
}
\value{
a list object with elements zero_n_check and farmer_practice
}
\description{
Reads shapefile of Zero Check Strip and buffers for Farmer practice
}
