% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_check_yield_potential.R
\name{zero_check_fertilized_yield_potential}
\alias{zero_check_fertilized_yield_potential}
\title{Calculate yield potential}
\usage{
zero_check_fertilized_yield_potential(
  yp_field,
  yp_farmer_practice,
  zero_check_strip,
  farmer_practice,
  slope,
  intercept
)
}
\description{
Calculate yield potential
}
