% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_check_variable_rate.R
\name{zero_check_variable_rate}
\alias{zero_check_variable_rate}
\title{Creates a variable rate nitrogen application image based on Nitrogen Rich Strip (n_rich_strip), Farmer/Target (farmer_practice), Growing Degree Days (GDD), and active_ndvi(Target).}
\usage{
zero_check_variable_rate(
  yld_fert_field,
  yld_ptnl_field,
  yld_ptnl_zero_check,
  Napp,
  grain_N = 0.024,
  NUE = 0.6
)
}
\arguments{
\item{yld_fert_field}{Unfertilized yield prediction for field (kg/ha)}

\item{yld_ptnl_field}{Fertilized yield prediction for field (kg/ha)}

\item{yld_ptnl_zero_check}{Fertilized yield prediction for field (kg/ha)}

\item{Napp}{fertilizer applied prior to sensing (kg/ha)}

\item{grain_N}{expected grain Nitrogen concentration (fraction)}

\item{NUE}{Nitrogen use efficiency (fraction)}
}
\value{
Variable nitrogen rate recommendation using the zero check strip algorithm
}
\description{
Creates a variable rate nitrogen application image based on Nitrogen Rich Strip (n_rich_strip), Farmer/Target (farmer_practice), Growing Degree Days (GDD), and active_ndvi(Target).
}
