% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_check_yield_potential.R
\name{zero_check_yield_potential}
\alias{zero_check_yield_potential}
\title{Calculate yield potential based on NDVI}
\usage{
zero_check_yield_potential(NDVI, GDD, A, b, YPmax = Inf)
}
\description{
Calculate yield potential based on NDVI
}
